requirements.txt
PyNaCl==1.4.0
discord.py==1.7.2
youtube_dl==2021.6.6
google-api-python-client==2.8.0
youtube-search-python==1.4.7 
